优化

+++ agentManage/agentManageAdmin-dubbo/src/main/java/com/ryx/credit/dao/order/OActivityMapper.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.Map;
 
 public interface OActivityMapper {
-    int countByExample(OActivityExample example);
+    long countByExample(OActivityExample example);
 
     int deleteByExample(OActivityExample example);
